pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "trackio"
description = "A lightweight, local-first, and free experiment tracking library built on top of Hugging Face Datasets and Spaces."
authors = [
    { name = "Abubakar Abid", email = "abubakar@huggingface.co" },
    { name = "Zach Nation", email = "zach@huggingface.co" },
]
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "pandas<3.0.0",
    "huggingface-hub<1.0.0",
    "gradio>=5.46.0,<6.0.0",
    "numpy<3.0.0",
    "pillow<12.0.0",
    "orjson>=3.0,<4.0.0",
]
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
]
dynamic = ["version"]

[project.urls]
homepage = "https://github.com/gradio-app/trackio"
repository = "https://github.com/gradio-app/trackio"

[project.optional-dependencies]
dev = ["pytest", "ruff==0.9.3", "pyarrow>=21.0"]
tensorboard = ["tbparse==0.0.9", "tensorboardX>=2.0.0,<3.0.0"]

[project.scripts]
trackio = "trackio.cli:main"

[tool.hatch.build.targets.wheel]
packages = ["trackio"]

[tool.hatch.version]
path = "trackio/package.json"
pattern = ".*\"version\":\\s*\"(?P<version>[^\"]+)\""

[tool.ruff]
# Exclude a variety of commonly ignored directories.
exclude = [
    ".git",
    ".mypy_cache",
    ".ruff_cache",
    ".venv",
    "__pycache__",
    "build",
    "dist",
]

# Same as Black.
line-length = 88
indent-width = 4

target-version = "py310"

[tool.ruff.lint]
# Enable Pyflakes (`F`) and a subset of the pycodestyle (`E`)
select = ["E", "F", "I"]
# Ignore line length violations
ignore = ["E501"]

# Allow autofix for all enabled rules (when `--fix`) is provided.
fixable = ["ALL"]

# Allow unused variables when underscore-prefixed.
dummy-variable-rgx = "^(_+|(_+[a-zA-Z0-9_]*[a-zA-Z0-9]+?))$"

[tool.ruff.format]
# Use single quotes for strings.
quote-style = "double"

# Indent with spaces, rather than tabs.
indent-style = "space"

# Like Black, respect magic trailing commas.
skip-magic-trailing-comma = false

# Like Black, automatically detect the appropriate line ending.
line-ending = "auto"
